z80asm2: separate location (filename and line number) from source text (read line and expanded line), so that Instr and Symbol can have Location objects telling where they were defined. Use a static StringTable to hold the filename, so that the Location object is just two ints.

+++ src/z80asm/z80asm2/memmap.h
@@ -22,20 +22,20 @@ public:
 
     void clear();
     const vector<uint8_t>& data() const { return m_data; }
-    size_t size() const { return m_data.size(); }
-    size_t pos() const;
-    void set_pos(size_t pos);
+    int size() const { return static_cast<int>(m_data.size()); }
+    int pos() const;
+    void set_pos(int pos);
 
     void align();
     void write_byte(uint8_t value);
     void write_short(uint16_t value);
     void write_long(uint32_t value);
-    void write_data(const uint8_t* data, size_t length);
+    void write_data(const uint8_t* data, int length);
 
     bool read_byte(uint8_t& value);
     bool read_short(uint16_t& value);
     bool read_long(uint32_t& value);
-    bool read_data(uint8_t* data, size_t length);
+    bool read_data(uint8_t* data, int length);
 
     bool write_file(const string& filename);
     bool read_file(const string& filename);
@@ -47,7 +47,7 @@ public:
 private:
     const static int ALIGN_SIZE = sizeof(uint32_t);    // alignment size in bytes
     vector<uint8_t> m_data;     // memory-mapped binary data
-    size_t m_pos{ 0 };          // read/write position in the data vector
+    int m_pos{ 0 };          // read/write position in the data vector
 
-    void extend_data(size_t count = 0);
+    void extend_data(int count = 0);
 };
